requirements.txt
yfinance==0.2.18
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
gspread==5.10.0
python-telegram-bot==20.4
ta==0.10.2
google-auth==2.22.0
google-auth-oauthlib==1.0.0
matplotlib==3.7.2
seaborn==0.12.2
